moved constants to /actions/
