--- src/redux/actionConstants.js
@@ -1,5 +1,0 @@
-export const SET_DATA = 'data/set_data'
-export const SET_LOADED = 'data/set_loaded'
-
-export const TOGGLE_PROVIDER = 'filters/toggle_provider'
-export const SET_CATEGORY = 'filters/set_category'
